// src/types/generated/patientAddRequestDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */
import type { Gender } from './gender';

export interface PatientAddRequestDto {
  /** @nullable */
  firstname?: string | null;
  /** @nullable */
  lastname?: string | null;
  /** @nullable */
  phone?: string | null;
  gender?: Gender;
  dateOfBirth?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  countryId?: number | null;
  /** @nullable */
  governorateId?: number | null;
  /** @nullable */
  districtId?: number | null;
  /** @nullable */
  localityId?: number | null;
  /** @nullable */
  emergencyContactName?: string | null;
  /** @nullable */
  emergencyContactPhone?: string | null;
  /** @nullable */
  bloodType?: string | null;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  chronicConditions?: string | null;
}