// src/types/generated/patientAddResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */

export interface PatientAddResponseDto {
  id?: string;
  /** @nullable */
  firstname?: string | null;
  /** @nullable */
  lastname?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt?: string;
}