// src/types/generated/patientEditResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Doctaura API
 * API for managing clinic operations, doctor schedules, and appointments
 * OpenAPI spec version: v1
 */

export interface PatientEditResponseDto {
  id?: string;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  updatedAt?: string;
}